requirements.txt
# development
asgiref==3.4.1
Django==3.2.6
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.14.0
python-dotenv==0.21.0
celery==5.2.2
redis==4.4.0
flower==1.2.0
petl==1.7.12
requests==2.28.1

# debug and testing
django-debug-toolbar==3.8.1
coverage>=6.5.0
black>=22.3.0
flake8>=4.0.1
autoflake>=2.0.0
mypy>=0.942
isort>=5.10.1
pytest>=7.1.1
pytest-django>=4.5.2
factory-boy==3.2.1